feat: mockup data display
