+++ app/dashboard/components/content.tsx
@@ -1,5 +1,5 @@
 import { Card, CardContent, CardDescription, CardFooter, CardHeader, CardTitle } from '@/components/ui/card'
-import { Table, TableBody, TableCell, TableHead, TableHeader, TableRow } from '@/components/ui/table'
+import { DashboardTable } from './table'
 
 function DashboardContent() {
   return (
@@ -30,36 +30,6 @@ function DashboardInfoCard() {
       <CardFooter>
         <p>Card Footer</p>
       </CardFooter>
-    </Card>
-  )
-}
-
-function DashboardTable() {
-  return (
-    <Card className="h-full p-0">
-      <Table>
-        <TableHeader>
-          <TableRow>
-            <TableHead className="w-[100px]">Invoice</TableHead>
-            <TableHead>Status</TableHead>
-            <TableHead>Method</TableHead>
-            <TableHead className="text-right">Amount</TableHead>
-          </TableRow>
-        </TableHeader>
-        <TableBody>
-          {Array.from({
-            length: 100,
-          }).map((_, i) => (
-            // eslint-disable-next-line react/no-array-index-key
-            <TableRow key={i}>
-              <TableCell className="font-medium">INV001</TableCell>
-              <TableCell>Paid</TableCell>
-              <TableCell>Credit Card</TableCell>
-              <TableCell className="text-right">.00</TableCell>
-            </TableRow>
-          ))}
-        </TableBody>
-      </Table>
     </Card>
   )
 }
